feat(tool): add unified rg tool replacing grep and glob (#44)
- New RgTool with two modes: content search and file listing
- files_only=true lists files matching glob pattern
- Default mode searches file contents with regex
- Removes old GrepTool and GlobTool
- Adds ripgrep as required dependency in README
