+++ packages/opencode/src/tool/registry.ts
@@ -1,8 +1,7 @@
 import { QuestionTool } from "./question"
 import { BashTool } from "./bash"
 import { EditTool } from "./edit"
-import { GlobTool } from "./glob"
-import { GrepTool } from "./grep"
+import { RgTool } from "./rg"
 import { BatchTool } from "./batch"
 import { ReadTool } from "./read"
 import { TaskTool } from "./task"
@@ -99,8 +98,7 @@ export namespace ToolRegistry {
       ...(["app", "cli", "desktop"].includes(Flag.OPENCODE_CLIENT) ? [QuestionTool] : []),
       BashTool,
       ReadTool,
-      GlobTool,
-      GrepTool,
+      RgTool,
       EditTool,
       WriteTool,
       TaskTool,
